fix: price Long -> Double 타입으로 변경 대공사 작업

+++ backend/funco-trade/src/main/java/com/found_404/funco/crypto/controller/CryptoController.java
@@ -21,7 +21,7 @@ public class CryptoController {
 	private final CryptoPrice cryptoPrice;
 
 	@GetMapping()
-	public ResponseEntity<Map<String, Long>> getCryptoPrice(@RequestParam List<String> tickers) {
+	public ResponseEntity<Map<String, Double>> getCryptoPrice(@RequestParam List<String> tickers) {
 		return ResponseEntity.ok(cryptoPrice.getTickerPriceMap(tickers));
 	}
