+++ backend/funco-trade/src/main/java/com/found_404/funco/trade/dto/TradeDto.java
@@ -15,7 +15,7 @@ public record TradeDto(
         TradeType tradeType,
         Double volume,
         Long orderCash,
-        Long price,
+        Double price,
         LocalDateTime tradeDate
 ) {
